requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
boto==2.34.0
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7
django-appconf==0.6
django-bootstrap3==5.0.3
django-ckeditor==4.4.6
django-compressor==1.4
django-debug-toolbar==1.3.0
django-extensions==1.3.6
django-gcm==1.0.6
django-grappelli==2.6.1
django-libsass==0.2
django-model-utils==2.2
django-rest-swagger==0.3.2
django-storages==1.1.8
django-tastypie==0.12.1
django-toolbelt==0.0.1
djangorestframework==3.0.2
drf-nested-routers==0.9.0
gunicorn==18.0
html5lib==0.99999
ipython==3.1.0
libsass==0.7.0
model-mommy==1.2.4
Pillow==2.4.0
psycopg2==2.5.2
pystache==0.5.4
python-dateutil==2.4.2
python-Levenshtein==0.12.0
python-mimeparse==0.1.4
pytz==2015.2
PyYAML==3.11
requests==2.2.1
six==1.6.1
sqlparse==0.1.15
static==1.0.2
static3==0.5.1
wheel==0.24.0
